Support maximum theshold for each type of faultlog
Support maximum theshold entry for CPER type and CrashDump type.

When the number of fault log entries exceed the maximum limit,
remove the entry and remove the file that stored the additional
data.

When the number of CPER/Crashdump entries exceed the maximum limit,
remove the file that stored the additional data.

Upstream-Status: Pending

Signed-off-by: Dung Cao <dung@os.amperecomputing.com>